ehq_back/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
pymysql==1.1.0
cryptography==41.0.5